:label: [rename] 업로드 위젯 패키지 위치를 home 안으로 변경
